Refactor complete...I think.
