+++ scripts/SystemGenerator/ICelestialBody.cs
@@ -8,7 +8,6 @@ public interface ICelestialBody
 {
     Guid Guid { get; }
     string Name { get; }
-    bool Generated { get; }
     Vector2 Sector { get; }
     Vector3 SectorLocation { get; }
