--- scripts/SystemGenerator/NoiseGenerator.cs.uid
@@ -1,1 +1,0 @@
-uid://b1pbl0t4pv0mj
