+++ scripts/SystemGenerator/Terrain/Terrain3D.cs
@@ -1,16 +1,52 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using Godot;
+using Godot.Collections;
 
 namespace Planets.SystemGenerator.Terrain;
 
-public partial class Terrain3D : MeshInstance3D
+[GlobalClass]
+public abstract partial class Terrain3D : MeshInstance3D
 {
-    // Called when the node enters the scene tree for the first time.
-    public override void _Ready()
+    [Export]
+    public TerrainColor Colors { get; set; }
+
+    // [Export(PropertyHint.NodeType, $"{nameof(ArrayMesh)}")]
+    // public new Mesh Mesh { get; set; }
+
+    [Export]
+    public ulong Seed { get; set; } = 69000;
+#if DEBUG
+    [Export]
+    public ImageTexture3D Before { get; protected set; }
+
+    [Export]
+    public ImageTexture3D After { get; protected set; }
+#endif
+
+    protected Terrain3D()
     {
     }
 
-    // Called every frame. 'delta' is the elapsed time since the previous frame.
-    public override void _Process(double delta)
-    {
-    }
+
+    public abstract void Generate(bool generateLods, ShaderMaterial shaderMaterial);
+
+    // public override void _ValidateProperty(Dictionary property)
+    // {
+    //     if (property["name"].AsStringName() != PropertyName.Mesh)
+    //     {
+    //         base._ValidateProperty(property);
+    //         return;
+    //     }
+    // 
+    //     PropertyUsageFlags propertyUsageFlags = property["usage"].As<PropertyUsageFlags>();
+    // 
+    //     if (_useGeneratedMesh)
+    //         propertyUsageFlags |= PropertyUsageFlags.ReadOnly;
+    //     else
+    //         propertyUsageFlags &= ~PropertyUsageFlags.ReadOnly;
+    // 
+    //     property["usage"] = (int)propertyUsageFlags;
+    //     base._ValidateProperty(property);
+    // }
 }
